# cmd/api/docs/swagger.yaml
definitions:
  domain.Admin:
    properties:
      admin_name:
        maxLength: 12
        minLength: 4
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        maxLength: 15
        minLength: 8
        type: string
    required:
    - password
    type: object
  requests.Addcount:
    properties:
      count:
        minimum: 1
        type: integer
      product_id:
        type: integer
    required:
    - product_id
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/login:
    post:
      consumes:
      - application/json
      description: login admin with details.
      operationId: LoginAdmin
      parameters:
      - description: Input Field
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/domain.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - Admin
  /admin/logout:
    post:
      description: Logout as a user exit from the ecommerce site
      operationId: AdminLogout
      responses:
        "200":
          description: success
        "400":
          description: failed
      summary: Adminlogout
      tags:
      - Admin
...
  /admin/signup:
    post:
      consumes:
      - application/json
      description: Save admin with details.
      operationId: SaveAdmin
      parameters:
      - description: Input Field
        in: body
        name: inputs
        required: true
        schema:
          $ref: '#/definitions/domain.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: SaveAdmin
      tags:
      - Admin
  /admin/unblock/{user_id}:
    patch:
      consumes:
      - application/json
      description: Admin can unblock a blocked user
      operationId: unblock-user
      parameters:
      - description: ID of the user to be unblocked
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can unblock a blocked user
      tags:
      - Admin
  /cart/Addcount:
    put:
      consumes:
      - application/json
      description: user can delete their cartitems by id
      operationId: Add-Qantity
      parameters:
      - description: Input Field
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.Addcount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can delete a category
      tags:
      - Cart
  /cart/viewcart:
    get:
      description: user can see all productItem that stored in cart
      operationId: Cart
      responses:
        "200":
          description: successfully got user cart items
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: faild to get cart items
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: api for get all cart item of user
      tags:
      - Cart
...
swagger: "2.0"